internal: move control from base-core to base
